[WIP]: Redesign the IL Validation process
This commit aims to redesign the ILValidator in order to improve the
general IL Validation process and lay the groundwork for future improvements.

The issues addressed are listed below:

1. Compartmentalize Rules

The rules are categorized based on the “context” required to check whether they are
valid or not.
This gives us the following types:

- Rules that require one to check the entire method (TR::ResolvedMethodSymbol)
  for a certain property (i.e. Soundness checks)
- Rules that only confine themselves to looking at a particular Block(TR::Block).
- Rules that only check whether a particular node (TR::Node) conforms to a certain
  expectation.

2. Decoupling existing Rules

A large part of which involves decoupling the `Soundness Checks` from the rest of the rules.
Also split up complicated rules into more manageable (somewhat atomic) chunks.

3. Give more control over `when` and `what` we are validating against.

At different points in the pipeline we can choose to `validate` the IL only against a certain
subset of all the available rules.

TODO : The state is passed to the aforementioned `validate` method (as an enum value) and depending
on that `state`, we decide which subset of all the available Rules we should check against
(the set of `ALL` available rules is specified during the creation of the ILValidator object).
[ This comment should be removed in the final commit message. ]

4. Make it easier to add new rules. Also provide a set of utility functions in order to
write said Rules.

Due to the nature of how of intertwined the above issues are, this was done as
one commit.

Related Issue: #1901, #566, #1971

Signed-off-by: Rwitaban (Ray) Banerjee <rayb@ca.ibm.com>