Connect the ILValidator to the existing compiler technology
This commit makes the ILValidator a part of OMR::Compilation.

The ILValidator is initialized right after IL generation.
And can then be used to Validate the IL at any point after that.

Essentially allowing "state aware" validation of the IL.
For example, we can choose to `validate` the IL using a certain set of Rules
right after IL Generation which in turn can then be different from the set of Rules
we choose to use immediately before going into Codegenration.

Signed-off-by: Rwitaban (Ray) Banerjee <rayb@ca.ibm.com>