Add an IL Validation Rule for checking if aiadd and aiuadd are valid on the target platform
The aiadd OpCode is only valid when the platform we are compiling
on is a 32 bit one.

This is because we expect binary operations to operate on fields of the same length.
However, in the OMR compiler IL, 'integer' sized opcodes are always 32 bits.
Meaning on a non-32 bit platform this ends up trying to invoke a binary operand
on arguments with different length, which is undefined.

[The same reasoning applies to aiuadd, the unsigned counterpart of aiadd]

Fixes: Issue #566

Signed-off-by: Rwitaban (Ray) Banerjee <rayb@ca.ibm.com>