[___WIP___]: Add the notion of ILValidationRules (.hpp)
[___WIP____] Post Initial PR ideas.

Tries to introduce the notion of individual validation options
for each rule. Also a lot of other "interesting stuff" which are
more like long term goals.

Note that this is more of a proof of concept than anywhere near a base
for an actual implementation.

A lot of this code needs extensive cleaning up.